Set NativeShare as static class
