+++ Assets/Native Share Demo Scene/Test.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3deecf3e86e09a4488d1075dc04d9fa1
+timeCreated: 1437497110
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
