+++ Assets/Plugins/NativeShare.cs
@@ -7,31 +7,9 @@ using System.IO;
  * https://github.com/ChrisMaire/unity-native-sharing
  */
 
-public class NativeShare : MonoBehaviour {
-	public string ScreenshotName = "screenshot.png";
+public static class NativeShare {
 
-    public void ShareScreenshotWithText(string text)
-    {
-        string screenShotPath = Application.persistentDataPath + "/" + ScreenshotName;
-        if(File.Exists(screenShotPath)) File.Delete(screenShotPath);
-
-        Application.CaptureScreenshot(ScreenshotName);
-
-        StartCoroutine(delayedShare(screenShotPath, text));
-    }
-
-    //CaptureScreenshot runs asynchronously, so you'll need to either capture the screenshot early and wait a fixed time
-    //for it to save, or set a unique image name and check if the file has been created yet before sharing.
-    IEnumerator delayedShare(string screenShotPath, string text)
-    {
-        while(!File.Exists(screenShotPath)) {
-    	    yield return new WaitForSeconds(.05f);
-        }
-        
-        Share(text, screenShotPath, "");
-    }
-
-	public void Share(string shareText, string imagePath, string url, string subject = "")
+	public static void Share(string shareText, string imagePath, string url, string subject = "")
 	{
 #if UNITY_ANDROID
 		AndroidJavaClass intentClass = new AndroidJavaClass("android.content.Intent");
